+++ tests/testthat/test-writing-to-file.R
@@ -7,19 +7,38 @@ paths <- c("testfiles/standard_case.xml", "broken_path.txt")
 # # prepare correct result -----
 # "tests/testthat/testfiles/standard_case.xml" %>%
 #   find_article() %>%
-#   write_csv("tests/testthat/testfiles/correct_meta_data.csv", col_names = F)
+#   write_csv("tests/testthat/testfiles/correct_meta_data.csv", col_names = T)
+# 
+# "tests/testthat/testfiles/standard_case.xml" %>%
+#   find_article() %>%
+#   write_csv("tests/testthat/testfiles/correct_meta_data_wo_cols.csv",
+#             col_names = FALSE)
 # # nolint end
 
 # tests ------
 test_that("writing correct results to file works", {
   temp_dir <- tempdir()
-  jstor_convert_to_file(paths, 1, paste0(temp_dir, "meta_data"), find_article)
+  jstor_convert_to_file(paths, 1, paste0(temp_dir, "/meta_data"), find_article,
+                        col_names = TRUE)
 
   expect_identical(read_csv("testfiles/correct_meta_data.csv",
-                            col_names = FALSE),
-                   read_csv(paste0(temp_dir, "meta_data-1.csv"),
-                            col_names = FALSE))
+                            col_names = TRUE),
+                   read_csv(paste0(temp_dir, "/meta_data-1.csv"),
+                            col_names = TRUE))
 
+  unlink(temp_dir)
+})
+
+test_that("not writing column names works", {
+  temp_dir <- tempdir()
+  jstor_convert_to_file(paths, 1, paste0(temp_dir, "/meta_data"), find_article,
+                        col_names = FALSE)
+  
+  expect_identical(read_csv("testfiles/correct_meta_data_wo_cols.csv",
+                            col_names = FALSE),
+                   read_csv(paste0(temp_dir, "/meta_data-1.csv"),
+                            col_names = FALSE))
+  
   unlink(temp_dir)
 })
 
@@ -58,15 +77,28 @@ test_that("on windows only single core is used", {
 
 
 
-test_that("import wrapper works", {
+test_that("import wrapper works with column names", {
   temp_dir <- tempdir()
   jstor_import(paths, out_file = "meta_data", out_path = temp_dir,
-                       .f = find_article)
+               .f = find_article, col_names = T)
 
   expect_identical(read_csv("testfiles/correct_meta_data.csv",
-                            col_names = FALSE),
-                   read_csv(paste0(temp_dir, "meta_data-1.csv"),
-                            col_names = FALSE))
+                            col_names = TRUE),
+                   read_csv(paste0(temp_dir, "/meta_data-1.csv"),
+                            col_names = TRUE))
 
+  unlink(temp_dir)
+})
+
+test_that("import wrapper works without column names", {
+  temp_dir <- tempdir()
+  jstor_import(paths, out_file = "meta_data", out_path = temp_dir,
+               .f = find_article, col_names = FALSE)
+  
+  expect_identical(read_csv("testfiles/correct_meta_data_wo_cols.csv",
+                            col_names = FALSE),
+                   read_csv(paste0(temp_dir, "/meta_data-1.csv"),
+                            col_names = FALSE))
+  
   unlink(temp_dir)
 })